Removing the current forgot password mechanism and make it simple (#370)
added new UI of forget password

added new confirmation screen
